INTERNAL: Remove AuthThread and do auth on IO Thread
